New README.md and INSTALL.md.
Also fix some python/cygwin issues and imported header line endings.

+++ clean.py
@@ -1,14 +1,21 @@
 #!/usr/bin/env python
 
-import os, sys
+import os, sys, shutil
 
-file = os.popen('git status --untracked-files --short')
+if __name__ == '__main__':
+    file = os.popen('git status -u --porcelain')
 
-for line in file:
-    item = line.split(' ')
-    if item[0] == '??':
-        path = ' '.join(item[1:]).rstrip()
-        print(path)
-        os.remove(path)
-
-file.close()
+    for line in file:
+        item = line.split(' ')
+        if item[0] == '??':
+            path = ' '.join(item[1:]).rstrip()
+            print(path)
+            try:
+                if os.path.isfile(path):
+                    os.remove(path)
+                if os.path.isdir(path):
+                    shutil.rmtree(path)
+            except OSError:
+                None
+                
+    file.close()
